Move `Examples/SwiftUI-Gallery/Actomaton-Gallery/Root` into `Sources/Root`
