+++ Sources/Root/ExampleList/Examples/StateDiagramExample.swift
