+++ Sources/Root/ExampleList/Examples/VideoDetectorExample.swift
